Try suggestion #1 from approved API proposal: Move implementation to MemoryMappedFiles. Expose the refs in FileSystem.AccessControl.
