+++ src/libraries/System.IO.FileSystem.AccessControl/src/System.IO.FileSystem.AccessControl.TypeForwards.cs
@@ -1,0 +1,6 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.MemoryMappedFiles.MemoryMappedFileAcl))]
+[assembly: System.Runtime.CompilerServices.TypeForwardedTo(typeof(System.IO.MemoryMappedFiles.MemoryMappedFileSecurity))]
